Reorganize and add nextstrain data source
We can pull some important metadata from nextstrain as well. This adds
some logic to parse that data and use it in the search indexing. It is
very basic and not completely used yet. Mostly this is just here to show
we can do it, and I will have to inspect the data in more detail later.